+++ CHANGELOG.md
@@ -23,8 +23,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   extracted from the `.doc` field of the alias definition if it is a table
   with `.doc` and `.definition` properties.
 
-* Added command `jj file edit` for editing files in any revision without
-  needing to change the working copy.
+* Added commands `jj file edit` and `jj file set` for editing files in any
+  revision without needing to change the working copy.
 
 ### Fixed bugs